Create Sample Data

Create sample data for testing the package functionality.

dep_sample_data(index)

Arguments

index

A character scalar or vector listing deprivation measures to return. These include the area deprivation index ("adi"), the gini coefficient ("gini"), two versions of the neighborhood deprivation index by Messer ("ndi_m") and Powell and Wiley ("ndi_pw"), and four versions of the social vulnerability index ("svi10", "svi14", "svi20", and "svi20s").

Value

A tibble containing the raw 2022 American Community Sruvey data for the given index. Each tibble will contain observations for the 115 counties in Missouri.

Examples

## load sample data
dep_sample_data(index = "ndi_m")
#> # A tibble: 115 × 41
#>    GEOID NAME   YEAR B06009_001E B06009_001M B06009_002E B06009_002M B17017_002E
#>    <chr> <chr> <dbl>       <dbl>       <dbl>       <dbl>       <dbl>       <dbl>
#>  1 29001 Adai…  2022       13868         203        1110         175        1960
#>  2 29003 Andr…  2022       12659         100         700         184         682
#>  3 29005 Atch…  2022        3857          27         360          71         294
#>  4 29007 Audr…  2022       17336          95        2289         395        1296
#>  5 29009 Barr…  2022       24415          76        3534         418        2069
#>  6 29011 Bart…  2022        7997          52        1069         195        1030
#>  7 29013 Bate…  2022       11069          52        1272         212        1190
#>  8 29015 Bent…  2022       15146          43        1875         294        1275
#>  9 29017 Boll…  2022        7616          71        1196         260         588
#> 10 29019 Boon…  2022      111010          81        5813         656       12170
#> # ℹ 105 more rows
#> # ℹ 33 more variables: B17017_002M <dbl>, B19001_002E <dbl>, B19001_002M <dbl>,
#> #   B19001_003E <dbl>, B19001_003M <dbl>, B19001_004E <dbl>, B19001_004M <dbl>,
#> #   B19001_005E <dbl>, B19001_005M <dbl>, B19001_006E <dbl>, B19001_006M <dbl>,
#> #   B19058_002E <dbl>, B19058_002M <dbl>, B25115_012E <dbl>, B25115_012M <dbl>,
#> #   B25115_025E <dbl>, B25115_025M <dbl>, C24030_002E <dbl>, C24030_002M <dbl>,
#> #   C24030_018E <dbl>, C24030_018M <dbl>, C24030_019E <dbl>, …