Calculate Deprivation Measures

Calculates various measures of deprivation.

get_deprivation(geography, variables, keep_subscales = FALSE,
    keep_components = FALSE, output = "tidy", year, state = NULL,
    county = NULL, geometry = FALSE, keep_geo_vars = FALSE, shift_geo = FALSE,
    units)

Arguments

geography

A character scalar; one of "state", "county", or "tract"

variables

A character scalar or vector listing deprivation measures to return. These include the gini coefficient (gini)...

keep_subscales

A logical scalar; if FALSE (defualt), subscales for SVI, SVIp, and CCVI will not be returned. If TRUE, these components will be returned.

keep_components

A logical salar; if FALSE (defualt), only the main output measures will be returned. If TRUE, all individual Census variables for each requested measure of deprivation will be returned. Be aware that this can create very large output tables.

output

A character scalar; one of "tidy" (long output) or "wide" depending on the type of data format you want.

year

A numeric between 2010 and 2020

state

A state name or FIPS code (may be either character or numeric as well as scalar or vector)

county

A county FIPS code or vector of FIPS codes

geometry

A logical salar; if FALSE (defualt), tabular data are returned. If TRUE, a sf class object will be returned. This will only work when output = "wide".

keep_geo_vars

A logical salar; if FALSE (defualt), a minimum number of spatial variables are returned. If TRUE, a full set of TIGER/Line variables are returned. This is only relevant if geometry = TRUE.

shift_geo

A logical scalar; if FALSE (defualt), Alaska and Hawaii will appear in their cartographically correct locations. If TRUE, Alaska and Hawaii will be shifted south to facilitate more straightforward national mapping.

units

A character scalar; either "mi" (miles) or "km" (kilometers)